Done with toggle changes.
