+++ client/src/enums.ts
@@ -1,0 +1,4 @@
+export enum Themes {
+  DARK = 'black',
+  LIGHT = 'white',
+}
